--- a/cpp/src/join/mixed_join_kernels_semi.cu
+++ b/cpp/src/join/mixed_join_kernels_semi.cu
@@ -41,12 +41,9 @@ __attribute__((visibility("hidden"))) __launch_bounds__(block_size) __global__
                        table_device_view build,
                        row_hash const hash_probe,
                        row_equality const equality_probe,
-                       join_kind const join_type,
                        cudf::detail::semi_map_type::device_view hash_table_view,
-                       size_type* join_output_l,
-                       cudf::ast::detail::expression_device_view device_expression_data,
-                       cudf::size_type const* join_result_offsets,
-                       bool const swap_tables)
+                       bool* raw_flagged,
+                       cudf::ast::detail::expression_device_view device_expression_data)
 {
   // Normally the casting of a shared memory array is used to create multiple
   // arrays of different types from the shared memory buffer, but here it is
@@ -60,7 +57,7 @@ __attribute__((visibility("hidden"))) __launch_bounds__(block_size) __global__
 
   cudf::size_type const left_num_rows  = left_table.num_rows();
   cudf::size_type const right_num_rows = right_table.num_rows();
-  auto const outer_num_rows            = (swap_tables ? right_num_rows : left_num_rows);
+  auto const outer_num_rows            = left_num_rows;
 
   cudf::size_type outer_row_index = threadIdx.x + blockIdx.x * block_size;
 
@@ -70,11 +67,10 @@ __attribute__((visibility("hidden"))) __launch_bounds__(block_size) __global__
   if (outer_row_index < outer_num_rows) {
     // Figure out the number of elements for this key.
     auto equality = single_expression_equality<has_nulls>{
-      evaluator, thread_intermediate_storage, swap_tables, equality_probe};
+      evaluator, thread_intermediate_storage, false, equality_probe};
 
-    if ((join_type == join_kind::LEFT_ANTI_JOIN) !=
-        (hash_table_view.contains(outer_row_index, hash_probe, equality))) {
-      *(join_output_l + join_result_offsets[outer_row_index]) = outer_row_index;
+    if ((hash_table_view.contains(outer_row_index, hash_probe, equality))) {
+      *(raw_flagged + outer_row_index) = true;
     }
   }
 }
@@ -86,12 +82,9 @@ template __global__ void mixed_join_semi<DEFAULT_JOIN_BLOCK_SIZE, true>(
   table_device_view build,
   row_hash const hash_probe,
   row_equality const equality_probe,
-  join_kind const join_type,
   cudf::detail::semi_map_type::device_view hash_table_view,
-  size_type* join_output_l,
-  cudf::ast::detail::expression_device_view device_expression_data,
-  cudf::size_type const* join_result_offsets,
-  bool const swap_tables);
+  bool* raw_flagged,
+  cudf::ast::detail::expression_device_view device_expression_data);
 
 template __global__ void mixed_join_semi<DEFAULT_JOIN_BLOCK_SIZE, false>(
   table_device_view left_table,
@@ -100,12 +93,9 @@ template __global__ void mixed_join_semi<DEFAULT_JOIN_BLOCK_SIZE, false>(
   table_device_view build,
   row_hash const hash_probe,
   row_equality const equality_probe,
-  join_kind const join_type,
   cudf::detail::semi_map_type::device_view hash_table_view,
-  size_type* join_output_l,
-  cudf::ast::detail::expression_device_view device_expression_data,
-  cudf::size_type const* join_result_offsets,
-  bool const swap_tables);
+  bool* raw_flagged,
+  cudf::ast::detail::expression_device_view device_expression_data);
 
 }  // namespace detail