--- a/cpp/src/join/mixed_join_semi.cu
+++ b/cpp/src/join/mixed_join_semi.cu
@@ -224,9 +224,8 @@ std::unique_ptr<rmm::device_uvector<size_type>> mixed_join_semi(
   auto const row_hash   = cudf::experimental::row::hash::row_hasher{preprocessed_probe};
   auto const hash_probe = row_hash.device_hasher(has_nulls);
 
-  // Flagged vector used to indicate indices from left/probe table which are present in output
-  auto flagged = std::make_unique<rmm::device_uvector<bool>>(probe.num_rows(), stream);
-  thrust::uninitialized_fill(rmm::exec_policy(stream), flagged->begin(), flagged->end(), false);
+  // Vector used to indicate indices from left/probe table which are present in output
+  auto left_table_keep_mask = rmm::device_uvector<bool>(probe.num_rows(), stream);
 
   if (has_nulls) {
     mixed_join_semi<DEFAULT_JOIN_BLOCK_SIZE, true>
@@ -238,7 +237,7 @@ std::unique_ptr<rmm::device_uvector<size_type>> mixed_join_semi(
         hash_probe,
         equality_probe,
         hash_table_view,
-        flagged->data(),
+        cudf::device_span<bool>(left_table_keep_mask),
         parser.device_expression_data);
   } else {
     mixed_join_semi<DEFAULT_JOIN_BLOCK_SIZE, false>
@@ -250,7 +249,7 @@ std::unique_ptr<rmm::device_uvector<size_type>> mixed_join_semi(
         hash_probe,
         equality_probe,
         hash_table_view,
-        flagged->data(),
+        cudf::device_span<bool>(left_table_keep_mask),
         parser.device_expression_data);
   }
 
@@ -261,10 +260,10 @@ std::unique_ptr<rmm::device_uvector<size_type>> mixed_join_semi(
     thrust::copy_if(rmm::exec_policy(stream),
                     thrust::counting_iterator<size_type>(0),
                     thrust::counting_iterator<size_type>(probe.num_rows()),
-                    flagged->begin(),
+                    left_table_keep_mask.begin(),
                     gather_map->begin(),
-                    [join_type] __device__(bool flag) {
-                      return flag == (join_type == detail::join_kind::LEFT_SEMI_JOIN);
+                    [join_type] __device__(bool keep_row) {
+                      return keep_row == (join_type == detail::join_kind::LEFT_SEMI_JOIN);
                     });
 
   gather_map->resize(thrust::distance(gather_map->begin(), gather_map_end), stream);