@wence- 2024-03-20
As a followup, which would be an API-break and a change from the normal (not semi-) joins, I wonder if it is worthwhile removing the `thrust::copy_if` that converts the boolean keep mask into a gather map. We could instead just return the mask directly and then compute the output of the merge with `apply_boolean_mask` rather than `gather`. This would reduce the memory footprint (since without computing the number of true values in the keep mask, we must first allocate space for `num_rows` integers and then compact after the `thrust::copy_if` call), and maybe perform better. However, it may just move the costs around a bit, so not necessarily an improvement.