@wence- 2024-03-21
> > As a followup, which would be an API-break and a change from the normal (not semi-) joins, I wonder if it is worthwhile removing the `thrust::copy_if` that converts the boolean keep mask into a gather map. We could instead just return the mask directly and then compute the output of the merge with `apply_boolean_mask` rather than `gather`. This would reduce the memory footprint (since without computing the number of true values in the keep mask, we must first allocate space for `num_rows` integers and then compact after the `thrust::copy_if` call), and maybe perform better. However, it may just move the costs around a bit, so not necessarily an improvement.
> 
> If this table is involved in another join operation, we would have to first materialize the rows numbers if I understand correctly. I am not sure if the performance difference would be worth the additional complexity.
> 
> For memory footprint, we could add a prefix sum kernel and only reserve the required memory, but this may not be a good tradeoff

Well, the gather map that is currently returned is just the sparse representation of the boolean mask. So you would move from doing:
```
gather_map = semi_join(left, right);
result = cudf::gather(left, gather_map);
```

To:
```
mask = semi_join(left, right);
result = cudf::apply_boolean_mask(left, mask);
```

That puts (in many cases) the responsibility of constructing the sparse gather map on `apply_boolean_mask` (which does run a prefix-sum kernel to not over-allocate), but devolves to `thrust::copy_if` directly from the mask into the output in cases where it is possible.